% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Append}
\alias{Append}
\title{Append a pattern to the end (Also aliased to Append)}
\usage{
Append(Regularity_object, ...)
}
\arguments{
\item{Regularity_object}{created from Regularity}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a Regularity object
}
\description{
Append a pattern to the end (Also aliased to Append)
}
\examples{
Regularity() \%>\% StartWith(3, 'digits') \%>\% Append('-')
}

