% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EndWith}
\alias{EndWith}
\title{The line must end with the specified pattern. Note that the regr
objext cannot be modifed once EndWith has been called on it}
\usage{
EndWith(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
The line must end with the specified pattern. Note that the regr
objext cannot be modifed once EndWith has been called on it
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% EndWith('>')
}

