% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Maybe}
\alias{Maybe}
\title{Zero or one of the specified pattern}
\usage{
Maybe(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
Zero or one of the specified pattern
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% Maybe('digit')
}

