% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OneOrMore}
\alias{OneOrMore}
\title{Specify that the pattern or identifer should appear one or many times}
\usage{
OneOrMore(Regularity_object, pattern)
}
\arguments{
\item{Regularity_object}{created from Regularity}

\item{pattern}{the pattern to be matched}
}
\value{
a Regularity object
}
\description{
Specify that the pattern or identifer should appear one or many times
}
\examples{
Regularity() \%>\% StartWith(3, 'digits') \%>\% OneOrMore('a')
}

