% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StartsWith}
\alias{StartsWith}
\title{The line must start with the specified pattern (aliased to StartWith function)}
\usage{
StartsWith(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
The line must start with the specified pattern (aliased to StartWith function)
}
\examples{
regr() \%>\% StartWith(3, 'digits')
}

