% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StartWith}
\alias{StartWith}
\title{The line must start with the specified pattern}
\usage{
StartWith(Regularity_object, ...)
}
\arguments{
\item{Regularity_object}{created from Regularity}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a Regularity object
}
\description{
The line must start with the specified pattern
}
\examples{
Regularity() \%>\% StartWith(3, 'digits')
}

