% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Then}
\alias{Then}
\title{Append a pattern to the end (Also aliased to then)}
\usage{
Then(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
Append a pattern to the end (Also aliased to then)
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% Then('-')
}

