% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ZeroOrMore}
\alias{ZeroOrMore}
\title{Specify that the pattern or identifer should appear zero or many times}
\usage{
ZeroOrMore(regr_object, pattern)
}
\arguments{
\item{regr_object}{created from regr}

\item{pattern}{the pattern to be matched}
}
\value{
a regr object
}
\description{
Specify that the pattern or identifer should appear zero or many times
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% OneOf('a')
}

