% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Append}
\alias{Append}
\title{Append a pattern to the end (Also aliased to Append)}
\usage{
Append(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
Append a pattern to the end (Also aliased to Append)
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% Append('-')
}

