% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Between}
\alias{Between}
\title{Specify a bounded repetition, e.g. between(c(2,4), 'digits')}
\usage{
Between(regr_object, range, pattern)
}
\arguments{
\item{regr_object}{created from regr}

\item{range}{vector containing the upper and lower bound e.g. c(1, 5)}

\item{pattern}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
Specify a bounded repetition, e.g. between(c(2,4), 'digits')
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% Between(c(1, 2), 'a')
}

