% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EndWith}
\alias{EndWith}
\title{The line must end with the specified pattern. Note that the Regularity
objext cannot be modifed once EndWith has been called on it}
\usage{
EndWith(Regularity_object, ...)
}
\arguments{
\item{Regularity_object}{created from Regularity}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a Regularity object
}
\description{
The line must end with the specified pattern. Note that the Regularity
objext cannot be modifed once EndWith has been called on it
}
\examples{
Regularity() \%>\% StartWith(3, 'digits') \%>\% EndWith('>')
}

