% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OneOf}
\alias{OneOf}
\title{Specify an alternation, e.g. one_of c('a', 'b', 'c')}
\usage{
OneOf(regr_object, pattern)
}
\arguments{
\item{regr_object}{created from regr}

\item{pattern}{the pattern to be matched}
}
\value{
a regr object
}
\description{
Specify an alternation, e.g. one_of c('a', 'b', 'c')
}
\examples{
regr() \%>\% StartWith(3, 'digits') \%>\% OneOf(c('a', 'b', 'c'))
}

