% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StartsWith}
\alias{StartsWith}
\title{The line must start with the specified pattern (aliased to StartWith function)}
\usage{
StartsWith(Regularity_object, ...)
}
\arguments{
\item{Regularity_object}{created from Regularity}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a Regularity object
}
\description{
The line must start with the specified pattern (aliased to StartWith function)
}
\examples{
Regularity() \%>\% StartWith(3, 'digits')
}

