% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StartWith}
\alias{StartWith}
\title{The line must start with the specified pattern}
\usage{
StartWith(regr_object, ...)
}
\arguments{
\item{regr_object}{created from regr}

\item{...}{can be either a regex pattern or count and pattern}
}
\value{
a regr object
}
\description{
The line must start with the specified pattern
}
\examples{
regr() \%>\% StartWith(3, 'digits')
}

